+++ docs/news.rst
@@ -24,6 +24,7 @@ Psycopg 3.1 (unreleased)
 Psycopg 3.0.9 (unreleased)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+- Set `Error.sqlstate` when an unknown code is received (:ticket:`#225`).
 - Add the `!tzdata` package as a dependency on Windows in order to handle time
   zones (:ticket:`#223`).